+++ app/components/primer/alpha/tab_nav.rb
@@ -13,11 +13,9 @@ module Primer
     #     accessibility considerations.
     class TabNav < Primer::Component
       include Primer::TabbedComponentHelper
+      include Primer::TabNavHelper
 
       status :alpha
-
-      DEFAULT_EXTRA_ALIGN = :left
-      EXTRA_ALIGN_OPTIONS = [DEFAULT_EXTRA_ALIGN, :right].freeze
 
       BODY_TAG_DEFAULT = :ul
 
@@ -29,11 +27,7 @@ module Primer
       # @param selected [Boolean] Whether the tab is selected.
       # @param system_arguments [Hash] <%= link_to_system_arguments_docs %>
       renders_many :tabs, lambda { |selected: false, **system_arguments|
-        system_arguments[:classes] = class_names(
-          "tabnav-tab",
-          system_arguments[:classes]
-        )
-
+        system_arguments[:classes] = tab_nav_tab_classes(system_arguments[:classes])
         Primer::Navigation::TabComponent.new(
           list: true,
           selected: selected,
@@ -44,8 +38,8 @@ module Primer
       # Renders extra content to the `TabNav`. This will be rendered after the tabs.
       #
       # @param align [Symbol] <%= one_of(Primer::Alpha::TabNav::EXTRA_ALIGN_OPTIONS) %>
-      renders_one :extra, lambda { |align: DEFAULT_EXTRA_ALIGN, &block|
-        @align = fetch_or_fallback(EXTRA_ALIGN_OPTIONS, align, DEFAULT_EXTRA_ALIGN)
+      renders_one :extra, lambda { |align: EXTRA_ALIGN_DEFAULT, &block|
+        @align = fetch_or_fallback(EXTRA_ALIGN_OPTIONS, align, EXTRA_ALIGN_DEFAULT)
 
         view_context.capture { block&.call }
       }
@@ -118,25 +112,18 @@ module Primer
       # @param body_arguments [Hash] <%= link_to_system_arguments_docs %> for the body wrapper.
       # @param system_arguments [Hash] <%= link_to_system_arguments_docs %>
       def initialize(label:, tag: TAG_DEFAULT, body_arguments: {}, **system_arguments)
-        @align = DEFAULT_EXTRA_ALIGN
-
+        @align = EXTRA_ALIGN_DEFAULT
         @system_arguments = system_arguments
         @body_arguments = body_arguments
 
         @system_arguments[:tag] = fetch_or_fallback(TAG_OPTIONS, tag, TAG_DEFAULT)
-        @system_arguments[:classes] = class_names(
-          "tabnav",
-          system_arguments[:classes]
-        )
+        @system_arguments[:classes] = tab_nav_classes(system_arguments[:classes])
 
         @body_arguments = body_arguments
         @body_arguments[:tag] = BODY_TAG_DEFAULT
-        @body_arguments[:classes] = class_names(
-          "tabnav-tabs",
-          body_arguments[:classes]
-        )
+        @body_arguments[:classes] = tab_nav_body_classes(system_arguments[:classes])
 
-        @system_arguments[:tag] == :nav ? @system_arguments[:"aria-label"] = label : @body_arguments[:"aria-label"] = label
+        aria_label_for_page_nav(label)
       end
     end
   end